refactor: add langbot.bot module
