+++ langbot/__init__.py
@@ -1,0 +1,2 @@
+__author__ = "EcmaXp <ecmaxp@ecmaxp.kr>"
+__version__ = "0.2"
